refactor: FigurePen abstract class를 두어 도형 펜의 중복 코드 제거
