+++ app/src/main/java/woowacourse/paint/view/model/pen/RectPen.kt
@@ -1,44 +1,15 @@
 package woowacourse.paint.view.model.pen
 
-import android.graphics.Paint
 import android.graphics.Path
 import woowacourse.paint.graphics.AutoSortedRect
 import woowacourse.paint.view.model.pen.ink.Ink
 
 class RectPen(
-    override val ink: Ink = Ink(
-        Paint(Paint.ANTI_ALIAS_FLAG).apply {
-            style = Paint.Style.FILL
-        },
-        Path()
-    ),
-    val onAddInk: (Ink) -> Unit = { _ -> }
-) : DrawablePen {
-    private var lastPointX: Float = 0F
-    private var lastPointY: Float = 0F
-
-    override fun startPaint(pointX: Float, pointY: Float) {
-        lastPointX = pointX
-        lastPointY = pointY
-    }
-
-    override fun movePaint(pointX: Float, pointY: Float) {
-        ink.path.reset()
+    override val onAddInk: (Ink) -> Unit = { _ -> }
+) : FigurePen(onAddInk = onAddInk) {
+    override fun draw(pointX: Float, pointY: Float) {
         ink.path.addRect(
             AutoSortedRect(lastPointX, lastPointY, pointX, pointY).toRectF(), Path.Direction.CCW
         )
-    }
-
-    override fun cacheCurrentPaint() {
-        if (!ink.path.isEmpty)
-            onAddInk(ink)
-        ink.path.reset()
-    }
-
-    override fun setStrokeWidth(strokeWidth: Float) {
-    }
-
-    override fun setColor(color: Int) {
-        ink.paint.color = color
     }
 }
